+++ gin.go
@@ -26,15 +26,27 @@ func AddGinHandlers(r *gin.Engine, d *sql.DB, l *logrus.Logger, v *validator.Val
 		base := &resource.Base{Logger: l, DB: d, Validate: v, Resource: &resources[i]}
 		name := fmt.Sprintf("/%s", strcase.KebabCase(resources[i].Table))
 		nameID := fmt.Sprintf("%s/:id", name)
-		r.POST(name, GinHandler(handlers.CreateHandler(base)))
-		r.GET(nameID, GinHandler(handlers.RetrieveHandler(base)))
-		r.PUT(name, GinHandler(handlers.UpdateHandler(base)))
-		r.DELETE(nameID, GinHandler(handlers.DeleteHandler(base)))
-		r.GET(name, GinHandler(handlers.SearchHandler(base)))
 
-		for _, belongsTo := range resources[i].BelongsToFields {
-			nameBelongsTo := fmt.Sprintf("/%s/:id%s", strcase.KebabCase(belongsTo.Table), name)
-			r.GET(nameBelongsTo, GinHandler(handlers.GetBelongsToHandler(base, belongsTo)))
+		if !resources[i].OmitCreateRoute {
+			r.POST(name, GinHandler(handlers.CreateHandler(base)))
+		}
+		if !resources[i].OmitRetrieveRoute {
+			r.GET(nameID, GinHandler(handlers.RetrieveHandler(base)))
+		}
+		if !resources[i].OmitUpdateRoute {
+			r.PUT(name, GinHandler(handlers.UpdateHandler(base)))
+		}
+		if !resources[i].OmitDeleteRoute {
+			r.DELETE(nameID, GinHandler(handlers.DeleteHandler(base)))
+		}
+		if !resources[i].OmitSearchRoute {
+			r.GET(name, GinHandler(handlers.SearchHandler(base)))
+		}
+		if !resources[i].OmitBelongsToRoutes {
+			for _, belongsTo := range resources[i].BelongsToFields {
+				nameBelongsTo := fmt.Sprintf("/%s/:id%s", strcase.KebabCase(belongsTo.Table), name)
+				r.GET(nameBelongsTo, GinHandler(handlers.GetBelongsToHandler(base, belongsTo)))
+			}
 		}
 	}
 }
